Movida FirebaseConfig a un archivo independiente
