+++ src/components/ItemDetailContainer.jsx
@@ -3,15 +3,14 @@ import { useParams } from 'react-router-dom'
 
 import ItemDetail from './ItemDetail'
 
-import { app } from '../App'
-import { doc, getDoc, getFirestore } from 'firebase/firestore'
+import db from '../services/firebase'
+import { doc, getDoc } from 'firebase/firestore'
 
 function ItemDetailContainer () {
   const [item, setItem] = useState(null)
 
   const { id } = useParams()
 
-  const db = getFirestore(app)
   const refDoc = doc(db, 'items', id)
 
   useEffect(() => {
